pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = "test rest-x authentication"
authors = ["bagussr <bagusssyamsu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.3.3"
flask-restx = "^1.1.0"
Flask-SQLAlchemy = "^3.0.5"
flask-migrate = "^4.0.4"
python-dotenv = "^1.0.0"
pymysql = "^1.1.0"
marshmallow = "^3.20.1"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
